==> notebooks/7. RF_model.ipynb <==
from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from pyspark.ml.feature import VectorAssembler, StringIndexer
from pyspark.ml.classification import RandomForestClassifier
from pyspark.ml import Pipeline
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from pyspark.mllib.evaluation import MulticlassMetrics


spark = (
 SparkSession.builder.appName("project 1 RF")
 .config("spark.sql.repl.eagerEval.enabled", True)
 .config("spark.executor.memory","6G")
 .config("spark.driver.memory","4G")
 .config("spark.sql.parquet.cacheMetadata", "true")
 .config("spark.sql.session.timeZone", "Etc/UTC")
 .config('spark.driver.maxResultSize', '2048m')
 .getOrCreate()
)

data = spark.read.parquet('../data/curated/merged_sdf.parquet')

data.where(F.col('tip_amount') >= 3).count() / data.count()

## Adjusting the columns to train the model ##

column = ['PULocationID','DOLocationID', 'Airport', 'Congestion', 'Weekend', 'tip']

def discretization(sdf):
    sdf = sdf.withColumn('tip', (F.col('tip_amount') >= 3).cast('BOOLEAN'))
    sdf = sdf.select(column)
    return sdf.withColumn('tip_numeric', F.col('tip').cast('int')).select(['PULocationID','DOLocationID', 'Airport', 'Congestion', 'Weekend', 'tip_numeric'])

data = discretization(data)

data.where(F.col('tip_numeric') != 0).count() / data.count()

## Model training ##

# Assuming 'label' is the target column and 'features' are the feature columns
assembler = VectorAssembler(inputCols=['PULocationID','DOLocationID', 'Airport', 'Congestion', 'Weekend'], outputCol="features")

# Takes 8 mins to run
rf = RandomForestClassifier(featuresCol="features", labelCol="tip_numeric", numTrees=100, maxDepth=5)

pipeline = Pipeline(stages=[assembler, rf])

train_data, test_data = data.randomSplit([0.8, 0.2], seed=42)

model = pipeline.fit(train_data)


## Evaluating the result ##

predictions = model.transform(test_data)

evaluator = MulticlassClassificationEvaluator(labelCol="tip_numeric", predictionCol="prediction", metricName="accuracy")
accuracy = evaluator.evaluate(predictions)
print("Accuracy:", accuracy)


prediction_and_label = predictions.select("prediction", "tip_numeric").rdd.map(lambda row: (float(row.prediction), float(row.tip_numeric)))

# Instantiate the MulticlassMetrics class
metrics = MulticlassMetrics(prediction_and_label)

# Compute the confusion matrix
confusion_matrix = metrics.confusionMatrix()


print(confusion_matrix)
